[Process] Add ProcessableInterface
